Aprimora a lógica de solicitação de corrida, ajusta o cálculo do valor mínimo e exibe o valor estimado ao solicitar corrida. Melhora a verificação de login do usuário para ser case insensitive.

+++ src/br/com/corrida/repository/CorridaRepository.java
@@ -26,7 +26,7 @@ public class CorridaRepository {
             while ((linha = reader.readLine()) != null) {
                 String[] partes = linha.split(";");
                 if (partes.length == 5) {
-                    Corrida corrida = new Corrida(null, Integer.parseInt(partes[3]));
+                    Corrida corrida = new Corrida(null, Double.parseDouble(partes[3]));
                     corrida.setId(Integer.parseInt(partes[0]));
                     corrida.setMotoqueiro(null);
                     // Note: Usuario and Motoqueiro should be set by their respective services
